+++ frontend/openchat-shared/src/domain/gif.ts
